requirements.txt
# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0

# Database
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
python-dateutil>=2.8.0
